app/requirements/base.txt
asgiref==3.4.1
Django==4.0.1
django-countries==7.2.1
django-localflavor==3.1
psycopg2-binary==2.8.6
python-stdnum==1.17
pytz==2021.3
sqlparse==0.4.2
